organization: move program-termination logic into separate file, add common.hpp for aggregating headers
