+++ src/common/logger.cpp
@@ -6,57 +6,12 @@
 
 #include "common/logger.hpp"
 
-#include <csignal>
-#include <cstdlib>
+#include "common/terminator.hpp"
 
-#include "axom/slic.hpp"
-#include "mpi.h"
-
-namespace {
-/**
- * The actual signal handler - must have this exact signature
- * MPI is probably dead by this point, SLIC is definitely dead
- */
-void signalHandler(int signal)
-{
-  std::cerr << "[SIGNAL]: Received signal " << signal << ", exiting" << std::endl;
-  serac::exitGracefully(true);
-}
-
-/**
- * Registers the signalHandler function to handle various fatal
- * signals
- * @note The behavior of MPI when a signal is sent to the mpirun
- * process is implementation-defined.  OpenMPI will send a SIGTERM
- * (which can be caught) and then a SIGKILL (which cannot) a few
- * seconds later.  PlatformMPI will propagate the signal to the
- * individual processes.  Before propagating or sending a signal,
- * MPI may or may not call MPI_Finalize/MPI_Abort.
- */
-void registerSignals()
-{
-  std::signal(SIGINT, signalHandler);
-  std::signal(SIGABRT, signalHandler);
-  std::signal(SIGSEGV, signalHandler);
-  std::signal(SIGTERM, signalHandler);
-}
-}  // namespace
+// #include "axom/slic.hpp"
+// #include "mpi.h"
 
 namespace serac {
-
-void exitGracefully(bool error)
-{
-  if (axom::slic::isInitialized()) {
-    serac::logger::flush();
-    serac::logger::finalize();
-  }
-  int mpi_finalized;
-  MPI_Finalized(&mpi_finalized);
-  if (!mpi_finalized) {
-    MPI_Finalize();
-  }
-  error ? exit(EXIT_FAILURE) : exit(EXIT_SUCCESS);
-}
 
 namespace logger {
 
@@ -68,7 +23,7 @@ bool initialize(MPI_Comm comm)
     slic::initialize();
   }
 
-  registerSignals();
+  terminator::registerSignals();
 
   int numRanks, rank;
   MPI_Comm_size(comm, &numRanks);
