+++ src/common/mesh_utils.cpp
@@ -7,6 +7,7 @@
 #include "mesh_utils.hpp"
 
 #include "common/logger.hpp"
+#include "common/terminator.hpp"
 #include "fmt/fmt.hpp"
 
 namespace serac {
